+++ Tests/CosmosDBMultiTenancy/CosmosDBMultiTenancy.Infrastructure/Repositories/CosmosDBRepositoryBase.cs
@@ -93,6 +93,21 @@ namespace CosmosDBMultiTenancy.Infrastructure.Repositories
             });
         }
 
+        public async Task<TDomain?> FindAsync(
+            Expression<Func<TDocumentInterface, bool>> filterExpression,
+            CancellationToken cancellationToken = default)
+        {
+            var documents = await _cosmosRepository.GetAsync(AdaptFilterPredicate(filterExpression), cancellationToken);
+
+            if (documents == null || !documents.Any())
+            {
+                return default;
+            }
+            var entity = LoadAndTrackDocument(documents.First());
+
+            return entity;
+        }
+
         public async Task<List<TDomain>> FindAllAsync(CancellationToken cancellationToken = default)
         {
             var documents = await _cosmosRepository.GetAsync(GetHasPartitionKeyExpression(_tenantId), cancellationToken);
@@ -119,7 +134,7 @@ namespace CosmosDBMultiTenancy.Infrastructure.Repositories
             }
         }
 
-        public virtual async Task<List<TDomain>> FindAllAsync(
+        public async Task<List<TDomain>> FindAllAsync(
             Expression<Func<TDocumentInterface, bool>> filterExpression,
             CancellationToken cancellationToken = default)
         {
@@ -132,7 +147,7 @@ namespace CosmosDBMultiTenancy.Infrastructure.Repositories
             return results;
         }
 
-        public virtual async Task<IPagedList<TDomain>> FindAllAsync(
+        public async Task<IPagedList<TDomain>> FindAllAsync(
             int pageNo,
             int pageSize,
             CancellationToken cancellationToken = default)
@@ -140,7 +155,7 @@ namespace CosmosDBMultiTenancy.Infrastructure.Repositories
             return await FindAllAsync(_ => true, pageNo, pageSize, cancellationToken);
         }
 
-        public virtual async Task<IPagedList<TDomain>> FindAllAsync(
+        public async Task<IPagedList<TDomain>> FindAllAsync(
             Expression<Func<TDocumentInterface, bool>> filterExpression,
             int pageNo,
             int pageSize,